+++ CHANGELOG.md
@@ -8,6 +8,11 @@ This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.htm
 
 ## [Unreleased]
 
+### Changed
+
+- Jaeger exporter helpers: added InstallNewPipeline and removed RegisterGlobal option instead. (#944)
+- Zipkin exporter helpers: pipeline methods introduced, new exporter method adjusted. (#944)
+
 ## [0.9.0] - 2020-07-20
 
 ### Added